Rework inclusion system
Inclusion system used to not work correctly before.
Loading of included resource used to happen only when you get it
from a context by hand. Thus data that was parsed using a scheme
and had an inclusion at deep levels (e.g. baseUriParameters ->
host -> description -> !include) raised an error, as an inclusion
was not processed.
Now all the included resources are loaded recursively at the start
of parsing process.