Fixed pathing and grabbing contents via network.
Changed the urllib2 import since urllib does not contain the urlopen
method.

Updated method for retrieving mime_type so it works in python 3.

Appended relative path with a '/', since urlparse.urljoin will remove
the dir in http://example.com/dir, instead of appending to it in the
case of example.com/dir/